refactor(ui): remove TabNavigator
