+++ src/components/AssetPickerList_deprecated.tsx
