+++ src/components/AssetPicker_deprecated.tsx
