+++ src/screens/Blocktank/OrderService_deprecated.tsx
@@ -2,7 +2,6 @@ import React, { ReactElement, useCallback, useEffect, useState } from 'react';
 import { StyleSheet, TouchableOpacity } from 'react-native';
 import { useSelector } from 'react-redux';
 import Clipboard from '@react-native-clipboard/clipboard';
-import type { StackScreenProps } from '@react-navigation/stack';
 
 import { Text, TextInput, View } from '../../styles/components';
 import NavigationHeader from '../../components/NavigationHeader';
@@ -17,11 +16,8 @@ import {
 import { claimChannel } from '../../store/actions/lightning';
 import Store from '../../store/types';
 import SafeAreaView from '../../components/SafeAreaView';
-import type { RootStackParamList } from '../../navigation/types';
 
-type Props = StackScreenProps<RootStackParamList, 'BlocktankOrder'>;
-
-const Order = (props: Props): ReactElement => {
+const Order = (props): ReactElement => {
 	const { navigation, route } = props;
 	const {
 		service: {
