+++ src/screens/Lightning/QuickConfirm.tsx
@@ -73,7 +73,7 @@ const QuickConfirm = ({
 			<NavigationHeader
 				title="Add Instant Payments"
 				onClosePress={(): void => {
-					navigation.navigate('Tabs');
+					navigation.navigate('Wallet');
 				}}
 			/>
 			<View style={styles.root}>
