+++ src/screens/Lightning/QuickSetup.tsx
@@ -151,7 +151,7 @@ const QuickSetup = ({
 			<NavigationHeader
 				title="Add Instant Payments"
 				onClosePress={(): void => {
-					navigation.navigate('Tabs');
+					navigation.navigate('Wallet');
 				}}
 			/>
 			<View style={styles.root}>
