+++ src/screens/Transfer/Confirm.tsx
@@ -104,7 +104,7 @@ const Confirm = ({
 			<NavigationHeader
 				title="Transfer Funds"
 				onClosePress={(): void => {
-					navigation.navigate('Tabs');
+					navigation.navigate('Wallet');
 				}}
 			/>
 			<View style={styles.root}>
